Disable XmPIXMAP mode for drag-and-drop
In XmPIXMAP mode, Motif will draw to the root window with
IncludeInferious without being under a GrabServer.

This results in a variety of graphical glitches. For example, when a
compositing window manager updates the screen it will erase the drag
icon. This continues until the icon leaves the origin window, when Motif
does a GrabServer. The GrabServer avoids the graphical glitches, but it
also freezes the rest of the desktop (which is impolite).

Given that all current X servers support SHAPE, and given that most
users are running a compositing window manager, XmDRAG_WINDOW is always
a better choice. Therefore, disable XmPIXMAP entirely.